+++ Kernel/Devices/MemoryDevice.h
@@ -39,7 +39,7 @@ public:
     MemoryDevice();
     ~MemoryDevice();
 
-    virtual KResultOr<Region*> mmap(Process&, FileDescription&, const Range&, size_t offset, int prot, bool shared) override;
+    virtual KResultOr<Region*> mmap(Process&, FileDescription&, const Range&, u64 offset, int prot, bool shared) override;
 
     // ^Device
     virtual mode_t required_mode() const override { return 0660; }
